Factor out expression codegen from function codegen (#155)
* please work

* moved back

* same trick

* move back

* factor expr codegen out

* add visit_, remove expr tests from func

* suggestiosn

* codegen test

* using generic visit

* extract method

* only stmt codegen

* reduce stop param

* fix compatability issues

* abc

* Visit an XYZ node.

* standard test method names

* suggestions
