+++ src/latexify/analyzers_test.py
@@ -150,3 +150,20 @@ def test_analyze_range_invalid(code: str) -> None:
         exceptions.LatexifySyntaxError, match=r"^Unsupported AST for analyze_range\.$"
     ):
         analyzers.analyze_range(node)
+
+
+@pytest.mark.parametrize(
+    "before,after",
+    [
+        ("n + 1", "n"),
+        ("n + 2", "n + 1"),
+        ("n - (-1)", "n - (-1) - 1"),
+        ("n - 1", "n - 2"),
+        ("1 * 2", "1 * 2 - 1"),
+    ],
+)
+def test_reduce_stop_parameter(before: str, after: str) -> None:
+    test_utils.assert_ast_equal(
+        analyzers.reduce_stop_parameter(ast_utils.parse_expr(before)),
+        ast_utils.parse_expr(after),
+    )
