+++ src/latexify/codegen/__init__.py
@@ -1,5 +1,6 @@
 """Package latexify.codegen."""
 
-from latexify.codegen import function_codegen
+from latexify.codegen import expression_codegen, function_codegen
 
+ExpressionCodegen = expression_codegen.ExpressionCodegen
 FunctionCodegen = function_codegen.FunctionCodegen
