+++ src/latexify/transformers/function_expander.py
@@ -27,7 +27,7 @@ class FunctionExpander(ast.NodeTransformer):
         self._functions = functions
 
     def visit_Call(self, node: ast.Call) -> ast.AST:
-        """Visitor of Call nodes."""
+        """Visit a Call node."""
         func_name = ast_utils.extract_function_name_or_none(node)
         if (
             func_name is not None
